+++ packages/@ember/-internals/glimmer/lib/register-curly-component.ts
@@ -1,0 +1,10 @@
+import { setInternalComponentManager } from '@glimmer/manager/lib/internal/api';
+
+import Component from './component';
+import { CURLY_COMPONENT_MANAGER } from './component-managers/curly';
+
+Component.reopenClass({
+  positionalParams: [],
+});
+
+setInternalComponentManager(CURLY_COMPONENT_MANAGER, Component);
