+++ packages/@ember/-internals/glimmer/lib/renderer.ts
@@ -1,9 +1,5 @@
-import { privatize as P } from '@ember/-internals/container';
-import { ENV } from '@ember/-internals/environment';
+import { ENV } from '@ember/-internals/environment/lib/env';
 import type { InternalOwner } from '@ember/-internals/owner';
-import { getOwner } from '@ember/-internals/owner';
-import { guidFor } from '@ember/-internals/utils';
-import { getViewElement, getViewId } from '@ember/-internals/views';
 import { assert } from '@ember/debug';
 import { _backburner, _getCurrentRunLoop } from '@ember/runloop';
 import {
@@ -15,91 +11,36 @@ import {
 } from '@glimmer/destroyable';
 import { DEBUG } from '@glimmer/env';
 import type {
-  Bounds,
+  ClassicResolver,
   Cursor,
   DebugRenderTree,
   Environment,
-  DynamicScope as GlimmerDynamicScope,
-  RenderResult as GlimmerRenderResult,
-  Template,
-  TemplateFactory,
   EvaluationContext,
-  CurriedComponent,
+  RenderResult as GlimmerRenderResult,
   TreeBuilder,
-  ClassicResolver,
 } from '@glimmer/interfaces';
+import { artifacts } from '@glimmer/program/lib/helpers';
+import { RuntimeOpImpl } from '@glimmer/program/lib/opcode';
+import { renderComponent as glimmerRenderComponent } from '@glimmer/runtime/lib/render';
+import { clientBuilder } from '@glimmer/runtime/lib/vm/element-builder';
+import { inTransaction, runtimeOptions } from '@glimmer/runtime/lib/environment';
+import { EvaluationContextImpl } from '@glimmer/opcode-compiler/lib/program-context';
+import { CURRENT_TAG, validateTag, valueForTag } from '@glimmer/validator/lib/validators';
+import type { SimpleDocument, SimpleElement } from '@simple-dom/interface';
 
-import type { Nullable } from '@ember/-internals/utility-types';
-import { artifacts, RuntimeOpImpl } from '@glimmer/program';
-import type { Reference } from '@glimmer/reference';
-import { createConstRef, UNDEFINED_REFERENCE, valueForRef } from '@glimmer/reference';
-import type { CurriedValue } from '@glimmer/runtime';
-import {
-  clientBuilder,
-  createCapturedArgs,
-  curry,
-  EMPTY_POSITIONAL,
-  inTransaction,
-  renderComponent as glimmerRenderComponent,
-  renderMain,
-  runtimeOptions,
-} from '@glimmer/runtime';
-import { dict } from '@glimmer/util';
-import { unwrapTemplate } from './component-managers/unwrap-template';
-import { CURRENT_TAG, validateTag, valueForTag } from '@glimmer/validator';
-import type { SimpleDocument, SimpleElement, SimpleNode } from '@simple-dom/interface';
-import RSVP from 'rsvp';
-import type Component from './component';
 import { hasDOM } from '../../browser-environment';
-import type ClassicComponent from './component';
-import { BOUNDS } from './component-managers/curly';
-import { createRootOutlet } from './component-managers/outlet';
-import { RootComponentDefinition } from './component-managers/root';
 import { EmberEnvironmentDelegate } from './environment';
 import ResolverImpl from './resolver';
-import type { OutletState } from './utils/outlet';
-import OutletView from './views/outlet';
-import { makeRouteTemplate } from './component-managers/route-template';
-import { EvaluationContextImpl } from '@glimmer/opcode-compiler';
 
 export type IBuilder = (env: Environment, cursor: Cursor) => TreeBuilder;
 
-export interface View {
-  parentView: Nullable<View>;
-  renderer: Renderer;
-  tagName: string | null;
-  elementId: string | null;
-  isDestroying: boolean;
-  isDestroyed: boolean;
-  [BOUNDS]: Bounds | null;
-}
-
-export class DynamicScope implements GlimmerDynamicScope {
-  constructor(
-    public view: View | null,
-    public outletState: Reference<OutletState | undefined>
-  ) {}
-
-  child() {
-    return new DynamicScope(this.view, this.outletState);
-  }
-
-  get(key: 'outletState'): Reference<OutletState | undefined> {
-    assert(
-      `Using \`-get-dynamic-scope\` is only supported for \`outletState\` (you used \`${key}\`).`,
-      key === 'outletState'
-    );
-    return this.outletState;
-  }
-
-  set(key: 'outletState', value: Reference<OutletState | undefined>) {
-    assert(
-      `Using \`-with-dynamic-scope\` is only supported for \`outletState\` (you used \`${key}\`).`,
-      key === 'outletState'
-    );
-    this.outletState = value;
-    return value;
-  }
+export interface RootState {
+  readonly type: 'classic' | 'component';
+  readonly destroyed: boolean;
+  readonly result: GlimmerRenderResult | undefined;
+  render(): void;
+  destroy(): void;
+  isFor(possibleRoot: unknown): boolean;
 }
 
 const NO_OP = () => {};
@@ -107,7 +48,7 @@ const NO_OP = () => {};
 // This wrapper logic prevents us from rerendering in case of a hard failure
 // during render. This prevents infinite revalidation type loops from occuring,
 // and ensures that errors are not swallowed by subsequent follow on failures.
-function errorLoopTransaction(fn: () => void) {
+export function errorLoopTransaction(fn: () => void) {
   if (DEBUG) {
     return () => {
       let didError = true;
@@ -133,9 +74,7 @@ function errorLoopTransaction(fn: () => void) {
   }
 }
 
-type RootState = ClassicRootState | ComponentRootState;
-
-class ComponentRootState {
+class ComponentRootState implements RootState {
   readonly type = 'component';
 
   #result: GlimmerRenderResult | undefined;
@@ -169,7 +108,7 @@ class ComponentRootState {
     });
   }
 
-  isFor(_component: ClassicComponent): boolean {
+  isFor(_root: unknown): boolean {
     return false;
   }
 
@@ -187,90 +126,6 @@ class ComponentRootState {
 
   get result(): GlimmerRenderResult | undefined {
     return this.#result;
-  }
-}
-
-class ClassicRootState {
-  readonly type = 'classic';
-  public id: string;
-  public result: GlimmerRenderResult | undefined;
-  public destroyed: boolean;
-  public render: () => void;
-  readonly env: Environment;
-
-  constructor(
-    public root: Component | OutletView,
-    context: EvaluationContext,
-    owner: object,
-    template: Template,
-    self: Reference<unknown>,
-    parentElement: SimpleElement,
-    dynamicScope: DynamicScope,
-    builder: IBuilder
-  ) {
-    assert(
-      `You cannot render \`${valueForRef(self)}\` without a template.`,
-      template !== undefined
-    );
-
-    this.id = root instanceof OutletView ? guidFor(root) : getViewId(root);
-    this.result = undefined;
-    this.destroyed = false;
-    this.env = context.env;
-
-    this.render = errorLoopTransaction(() => {
-      let layout = unwrapTemplate(template).asLayout();
-
-      let iterator = renderMain(
-        context,
-        owner,
-        self,
-        builder(context.env, { element: parentElement, nextSibling: null }),
-        layout,
-        dynamicScope
-      );
-
-      let result = (this.result = iterator.sync());
-
-      associateDestroyableChild(this, result);
-
-      this.render = errorLoopTransaction(() => {
-        if (isDestroying(result) || isDestroyed(result)) return;
-
-        return result.rerender({
-          alwaysRevalidate: false,
-        });
-      });
-    });
-  }
-
-  isFor(possibleRoot: unknown): boolean {
-    return this.root === possibleRoot;
-  }
-
-  destroy() {
-    let { result, env } = this;
-
-    this.destroyed = true;
-
-    this.root = null as any;
-    this.result = undefined;
-    this.render = undefined as any;
-
-    if (result !== undefined) {
-      /*
-       Handles these scenarios:
-
-       * When roots are removed during standard rendering process, a transaction exists already
-         `.begin()` / `.commit()` are not needed.
-       * When roots are being destroyed manually (`component.append(); component.destroy() case), no
-         transaction exists already.
-       * When roots are being destroyed during `Renderer#destroy`, no transaction exists
-
-       */
-
-      inTransaction(env, () => destroy(result!));
-    }
   }
 }
 
@@ -297,7 +152,17 @@ function loopBegin(): void {
   }
 }
 
-let renderSettledDeferred: RSVP.Deferred<void> | null = null;
+type Deferred = { promise: Promise<void>; resolve: () => void };
+
+function defer(): Deferred {
+  let resolve!: () => void;
+  let promise = new Promise<void>((r) => {
+    resolve = r;
+  });
+  return { promise, resolve };
+}
+
+let renderSettledDeferred: Deferred | null = null;
 /*
   Returns a promise which will resolve when rendering has settled. Settled in
   this context is defined as when all of the tags in use are "current" (e.g.
@@ -309,7 +174,7 @@ let renderSettledDeferred: RSVP.Deferred<void> | null = null;
 */
 export function renderSettled() {
   if (renderSettledDeferred === null) {
-    renderSettledDeferred = RSVP.defer();
+    renderSettledDeferred = defer();
     // if there is no current runloop, the promise created above will not have
     // a chance to resolve (because its resolved in backburner's "end" event)
     if (!_getCurrentRunLoop()) {
@@ -350,10 +215,6 @@ function loopEnd() {
 
 _backburner.on('begin', loopBegin);
 _backburner.on('end', loopEnd);
-
-interface ViewRegistry {
-  [viewId: string]: unknown;
-}
 
 type Resolver = ClassicResolver;
 
@@ -717,12 +578,12 @@ export function renderComponent(
 const RENDER_CACHE = new WeakMap<IntoTarget, RenderCacheEntry>();
 const RENDERER_CACHE = new WeakMap<object, BaseRenderer>();
 
-class BaseRenderer {
+export class BaseRenderer {
   static strict(
     owner: object,
     document: SimpleDocument | Document,
     options: { isInteractive: boolean; hasDOM?: boolean }
-  ) {
+  ): BaseRenderer {
     return new BaseRenderer(
       owner,
       { hasDOM: hasDOM, ...options },
@@ -801,208 +662,6 @@ class BaseRenderer {
   rerender(): void {
     this.state.scheduleRevalidate(this);
   }
-
-  // render(component: Component, options: { into: Cursor; args?: Record<string, unknown> }): void {
-  //   this.state.renderRoot(component);
-  // }
 }
 
-export class Renderer extends BaseRenderer {
-  static strict(
-    owner: object,
-    document: SimpleDocument | Document,
-    options: { isInteractive: boolean; hasDOM?: boolean }
-  ): BaseRenderer {
-    return new BaseRenderer(
-      owner,
-      { hasDOM: hasDOM, ...options },
-      document as SimpleDocument,
-      new ResolverImpl(),
-      clientBuilder
-    );
-  }
-
-  private _rootTemplate: Template;
-  private _viewRegistry: ViewRegistry;
-
-  static create(props: { _viewRegistry: any }): Renderer {
-    let { _viewRegistry } = props;
-    let owner = getOwner(props);
-    assert('Renderer is unexpectedly missing an owner', owner);
-    let document = owner.lookup('service:-document') as SimpleDocument;
-    let env = owner.lookup('-environment:main') as {
-      isInteractive: boolean;
-      hasDOM: boolean;
-    };
-    let rootTemplate = owner.lookup(P`template:-root`) as TemplateFactory;
-    let builder = owner.lookup('service:-dom-builder') as IBuilder;
-    return new this(owner, document, env, rootTemplate, _viewRegistry, builder);
-  }
-
-  constructor(
-    owner: InternalOwner,
-    document: SimpleDocument,
-    env: { isInteractive: boolean; hasDOM: boolean },
-    rootTemplate: TemplateFactory,
-    viewRegistry: ViewRegistry,
-    builder = clientBuilder,
-    resolver = new ResolverImpl()
-  ) {
-    super(owner, env, document, resolver, builder);
-    this._rootTemplate = rootTemplate(owner);
-    this._viewRegistry = viewRegistry || owner.lookup('-view-registry:main');
-  }
-
-  // renderer HOOKS
-
-  appendOutletView(view: OutletView, target: SimpleElement): void {
-    // TODO: This bypasses the {{outlet}} syntax so logically duplicates
-    // some of the set up code. Since this is all internal (or is it?),
-    // we can refactor this to do something more direct/less convoluted
-    // and with less setup, but get it working first
-    let outlet = createRootOutlet(view);
-    let { name, /* controller, */ template } = view.state;
-
-    let named = dict<Reference>();
-
-    named['Component'] = createConstRef(
-      makeRouteTemplate(view.owner, name, template as Template),
-      '@Component'
-    );
-
-    // TODO: is this guaranteed to be undefined? It seems to be the
-    // case in the `OutletView` class. Investigate how much that class
-    // exists as an internal implementation detail only, or if it was
-    // used outside of core. As far as I can tell, test-helpers uses
-    // it but only for `setOutletState`.
-    // named['controller'] = createConstRef(controller, '@controller');
-    // Update: at least according to the debug render tree tests, we
-    // appear to always expect this to be undefined. Not a definitive
-    // source by any means, but is useful evidence
-    named['controller'] = UNDEFINED_REFERENCE;
-    named['model'] = UNDEFINED_REFERENCE;
-
-    let args = createCapturedArgs(named, EMPTY_POSITIONAL);
-
-    this._appendDefinition(
-      view,
-      curry(0 as CurriedComponent, outlet, view.owner, args, true),
-      target
-    );
-  }
-
-  appendTo(view: ClassicComponent, target: SimpleElement): void {
-    let definition = new RootComponentDefinition(view);
-    this._appendDefinition(
-      view,
-      curry(0 as CurriedComponent, definition, this.state.owner, null, true),
-      target
-    );
-  }
-
-  _appendDefinition(
-    root: OutletView | ClassicComponent,
-    definition: CurriedValue,
-    target: SimpleElement
-  ): void {
-    let self = createConstRef(definition, 'this');
-    let dynamicScope = new DynamicScope(null, UNDEFINED_REFERENCE);
-    let rootState = new ClassicRootState(
-      root,
-      this.state.context,
-      this.state.owner,
-      this._rootTemplate,
-      self,
-      target,
-      dynamicScope,
-      this.state.builder
-    );
-    this.state.renderRoot(rootState, this);
-  }
-
-  cleanupRootFor(component: ClassicComponent): void {
-    // no need to cleanup roots if we have already been destroyed
-    if (isDestroyed(this)) {
-      return;
-    }
-
-    let roots = this.state.roots;
-
-    // traverse in reverse so we can remove items
-    // without mucking up the index
-    let i = roots.length;
-    while (i--) {
-      let root = roots[i];
-      assert('has root', root);
-      if (root.type === 'classic' && root.isFor(component)) {
-        root.destroy();
-        roots.splice(i, 1);
-      }
-    }
-  }
-
-  remove(view: ClassicComponent): void {
-    view._transitionTo('destroying');
-
-    this.cleanupRootFor(view);
-
-    if (this.state.isInteractive) {
-      view.trigger('didDestroyElement');
-    }
-  }
-
-  get _roots() {
-    return this.state.debug.roots;
-  }
-
-  get _inRenderTransaction() {
-    return this.state.debug.inRenderTransaction;
-  }
-
-  get _isInteractive() {
-    return this.state.debug.isInteractive;
-  }
-
-  get _context() {
-    return this.state.context;
-  }
-
-  register(view: any): void {
-    let id = getViewId(view);
-    assert(
-      'Attempted to register a view with an id already in use: ' + id,
-      !this._viewRegistry[id]
-    );
-    this._viewRegistry[id] = view;
-  }
-
-  unregister(view: any): void {
-    delete this._viewRegistry[getViewId(view)];
-  }
-
-  getElement(component: View): Nullable<Element> {
-    if (this._isInteractive) {
-      return getViewElement(component);
-    } else {
-      throw new Error(
-        'Accessing `this.element` is not allowed in non-interactive environments (such as FastBoot).'
-      );
-    }
-  }
-
-  getBounds(component: View): {
-    parentElement: SimpleElement;
-    firstNode: SimpleNode;
-    lastNode: SimpleNode;
-  } {
-    let bounds: Bounds | null = component[BOUNDS];
-
-    assert('object passed to getBounds must have the BOUNDS symbol as a property', bounds);
-
-    let parentElement = bounds.parentElement();
-    let firstNode = bounds.firstNode();
-    let lastNode = bounds.lastNode();
-
-    return { parentElement, firstNode, lastNode };
-  }
-}
+export type { DynamicScope, Renderer, View } from './classic-renderer';
