+++ packages/@ember/-internals/glimmer/lib/resolver.ts
@@ -1,7 +1,7 @@
 import type { InternalFactory, InternalOwner } from '@ember/-internals/owner';
 import { isFactory } from '@ember/-internals/owner';
 import { assert } from '@ember/debug';
-import { _instrumentStart } from '@ember/instrumentation';
+import { _instrumentStart } from '@ember/instrumentation/lib/internal-instrument';
 import { DEBUG } from '@glimmer/env';
 import type {
   ClassicResolver,
@@ -12,23 +12,19 @@ import type {
   TemplateFactory,
 } from '@glimmer/interfaces';
 import type { Nullable } from '@ember/-internals/utility-types';
+import { getComponentTemplate } from '@glimmer/manager/lib/public/template';
+import { getInternalComponentManager } from '@glimmer/manager/lib/internal/api';
+import { array } from '@glimmer/runtime/lib/helpers/array';
+import { concat } from '@glimmer/runtime/lib/helpers/concat';
+import { fn } from '@glimmer/runtime/lib/helpers/fn';
+import { get } from '@glimmer/runtime/lib/helpers/get';
+import { hash } from '@glimmer/runtime/lib/helpers/hash';
+import { on } from '@glimmer/runtime/lib/modifiers/on';
 import {
-  getComponentTemplate,
-  getInternalComponentManager,
-  setInternalHelperManager,
-} from '@glimmer/manager';
-import {
-  array,
-  concat,
-  fn,
-  get,
-  hash,
-  on,
   templateOnlyComponent,
   TEMPLATE_ONLY_COMPONENT_MANAGER,
-} from '@glimmer/runtime';
-import { isCurlyManager } from './component-managers/curly';
-import { CLASSIC_HELPER_MANAGER, isClassicHelper } from './helper';
+} from '@glimmer/runtime/lib/component/template-only';
+import { isCurlyManager } from './component-managers/curly-symbols';
 import { default as disallowDynamicResolution } from './helpers/-disallow-dynamic-resolution';
 import { default as inElementNullCheckHelper } from './helpers/-in-element-null-check';
 import { default as normalizeClassHelper } from './helpers/-normalize-class';
@@ -39,9 +35,6 @@ import { default as mut } from './helpers/mut';
 import { default as readonly } from './helpers/readonly';
 import { default as unbound } from './helpers/unbound';
 import { default as uniqueId } from './helpers/unique-id';
-
-import { mountHelper } from './syntax/mount';
-import { outletHelper } from './syntax/outlet';
 
 function instrumentationPayload(name: string) {
   return { object: `component:${name}` };
@@ -96,10 +89,13 @@ const BUILTIN_KEYWORD_HELPERS: Record<string, object> = {
   '-normalize-class': normalizeClassHelper,
   '-resolve': resolve,
   '-track-array': trackArray,
-  '-mount': mountHelper,
-  '-outlet': outletHelper,
   '-in-el-null': inElementNullCheckHelper,
 };
+
+export function registerBuiltInKeywordHelper(name: string, helper: object): void {
+  BUILTIN_KEYWORD_HELPERS[name] = helper;
+  BUILTIN_HELPERS[name] = helper;
+}
 
 const BUILTIN_HELPERS: Record<string, object> = {
   ...BUILTIN_KEYWORD_HELPERS,
@@ -134,7 +130,17 @@ const BUILTIN_MODIFIERS: Record<string, object> = {
   on,
 };
 
-const CLASSIC_HELPER_MANAGER_ASSOCIATED = new WeakSet();
+type ClassicHelperHandler = (
+  definition: object,
+
+  factory: any
+) => HelperDefinitionState | null;
+
+let classicHelperHandler: ClassicHelperHandler | null = null;
+
+export function registerClassicHelperHandler(handler: ClassicHelperHandler): void {
+  classicHelperHandler = handler;
+}
 
 export default class ResolverImpl implements ClassicResolver<InternalOwner> {
   private componentDefinitionCache: Map<object, ResolvedComponentDefinition | null> = new Map();
@@ -166,23 +172,11 @@ export default class ResolverImpl implements ClassicResolver<InternalOwner> {
       return null;
     }
 
-    if (typeof definition === 'function' && isClassicHelper(definition)) {
-      // For classic class based helpers, we need to pass the factoryFor result itself rather
-      // than the raw value (`factoryFor(...).class`). This is because injections are already
-      // bound in the factoryFor result, including type-based injections
-
-      if (DEBUG) {
-        // In DEBUG we need to only set the associated value once, otherwise
-        // we'll trigger an assertion
-        if (!CLASSIC_HELPER_MANAGER_ASSOCIATED.has(factory)) {
-          CLASSIC_HELPER_MANAGER_ASSOCIATED.add(factory);
-          setInternalHelperManager(CLASSIC_HELPER_MANAGER, factory);
-        }
-      } else {
-        setInternalHelperManager(CLASSIC_HELPER_MANAGER, factory);
+    if (typeof definition === 'function' && classicHelperHandler !== null) {
+      let result = classicHelperHandler(definition, factory);
+      if (result !== null) {
+        return result;
       }
-
-      return factory;
     }
 
     return definition;
