+++ packages/@ember/-internals/glimmer/lib/utils/to-bool.ts
@@ -1,15 +1,20 @@
 import { isHTMLSafe } from './string';
-import { get, tagForProperty } from '@ember/-internals/metal';
-import { isArray } from '@ember/array';
-import { isProxy } from '@ember/-internals/utils';
-import { consumeTag } from '@glimmer/validator';
+import { get } from '@ember/-internals/metal/lib/property_get';
+import { tagForProperty } from '@ember/-internals/metal/lib/tags';
+import { isEmberArray } from '@ember/array/-internals';
+import { isProxy } from '@ember/-internals/utils/lib/is_proxy';
+import { consumeTag } from '@glimmer/validator/lib/tracking';
+
+function isArrayLike(obj: unknown): obj is ArrayLike<unknown> {
+  return Array.isArray(obj) || isEmberArray(obj);
+}
 
 export default function toBool(predicate: unknown): boolean {
   if (isProxy(predicate)) {
     consumeTag(tagForProperty(predicate, 'content'));
 
     return Boolean(get(predicate, 'isTruthy'));
-  } else if (isArray(predicate)) {
+  } else if (isArrayLike(predicate)) {
     consumeTag(tagForProperty(predicate as object, '[]'));
 
     return (predicate as { length: number }).length !== 0;
