+++ packages/@ember/-internals/metal/lib/observer.ts
@@ -1,12 +1,14 @@
-import { ENV } from '@ember/-internals/environment';
-import { peekMeta } from '@ember/-internals/meta';
-import type { schedule } from '@ember/runloop';
+import { ENV } from '@ember/-internals/environment/lib/env';
+import { peekMeta } from '@ember/-internals/meta/lib/meta';
+import { registerAsyncObserverFlush, type schedule } from '@ember/runloop';
 import { registerDestructor } from '@glimmer/destroyable';
-import type { Tag } from '@glimmer/validator';
-import { CURRENT_TAG, tagMetaFor, validateTag, valueForTag } from '@glimmer/validator';
+import type { Tag } from '@glimmer/interfaces';
+import { CURRENT_TAG, validateTag, valueForTag } from '@glimmer/validator/lib/validators';
+import { tagMetaFor } from '@glimmer/validator/lib/meta';
 import { getChainTagsForKey } from './chain-tags';
 import changeEvent from './change_event';
 import { addListener, removeListener, sendEvent } from './events';
+import { registerObserverDeactivationHooks, registerObserverFlushSync } from './property_events';
 
 interface ActiveObserver {
   tag: Tag;
@@ -269,3 +271,7 @@ function destroyObservers(target: object) {
   if (SYNC_OBSERVERS.size > 0) SYNC_OBSERVERS.delete(target);
   if (ASYNC_OBSERVERS.size > 0) ASYNC_OBSERVERS.delete(target);
 }
+
+registerObserverFlushSync(flushSyncObservers);
+registerObserverDeactivationHooks(suspendedObserverDeactivation, resumeObserverDeactivation);
+registerAsyncObserverFlush(flushAsyncObservers);
