+++ packages/@ember/-internals/metal/lib/property_set.ts
@@ -1,10 +1,18 @@
 import { lookupDescriptor, setWithMandatorySetter, toString } from '@ember/-internals/utils';
 import { assert } from '@ember/debug';
 import { DEBUG } from '@glimmer/env';
-import { COMPUTED_SETTERS } from './decorator';
 import { isPath } from './path_cache';
 import { notifyPropertyChange } from './property_events';
 import { getPossibleMandatoryProxyValue, _getPath as getPath } from './property_get';
+
+let computedSetterCheck: ((setter: (this: object, value: unknown) => void) => boolean) | null =
+  null;
+
+export function registerComputedSetterCheck(
+  fn: (setter: (this: object, value: unknown) => void) => boolean
+): void {
+  computedSetterCheck = fn;
+}
 
 interface ExtendedObject {
   isDestroyed?: boolean;
@@ -68,7 +76,12 @@ export function set<T>(obj: object, keyName: string, value: T, tolerant?: boolea
 export function _setProp(obj: object, keyName: string, value: any) {
   let descriptor = lookupDescriptor(obj, keyName);
 
-  if (descriptor !== null && COMPUTED_SETTERS.has(descriptor.set!)) {
+  if (
+    descriptor !== null &&
+    descriptor.set !== undefined &&
+    computedSetterCheck !== null &&
+    computedSetterCheck(descriptor.set)
+  ) {
     (obj as any)[keyName] = value;
     return value;
   }
