+++ packages/@ember/runloop/index.ts
@@ -1,8 +1,15 @@
 import { assert } from '@ember/debug';
 import { onErrorTarget } from '@ember/-internals/error-handling';
-import { flushAsyncObservers } from '@ember/-internals/metal';
 import Backburner, { type Timer, type DeferredActionQueues } from 'backburner.js';
 import type { AnyFn } from '@ember/-internals/utility-types';
+
+let asyncObserverFlush: ((schedule: AsyncObserverScheduler) => void) | null = null;
+
+type AsyncObserverScheduler = typeof schedule;
+
+export function registerAsyncObserverFlush(fn: (schedule: AsyncObserverScheduler) => void): void {
+  asyncObserverFlush = fn;
+}
 
 export type { Timer };
 
@@ -45,12 +52,16 @@ function onBegin(current: DeferredActionQueues) {
 function onEnd(_current: DeferredActionQueues, next: DeferredActionQueues) {
   currentRunLoop = next;
 
-  flushAsyncObservers(schedule);
+  if (asyncObserverFlush !== null) {
+    asyncObserverFlush(schedule);
+  }
 }
 
 function flush(queueName: string, next: () => void) {
   if (queueName === 'render' || queueName === _rsvpErrorQueue) {
-    flushAsyncObservers(schedule);
+    if (asyncObserverFlush !== null) {
+      asyncObserverFlush(schedule);
+    }
   }
 
   next();
