+++ packages/@glimmer/runtime/lib/get-debug-name.ts
@@ -1,0 +1,8 @@
+import type { ComponentDefinition } from '@glimmer/interfaces';
+
+export function getDebugName(
+  definition: ComponentDefinition,
+  manager = definition.manager
+): string {
+  return definition.resolvedName ?? definition.debugName ?? manager.getDebugName(definition.state);
+}
